UPSTREAM: HID: sony: Fix input device leak when connecting a DS4 twice using USB/BT
When a user connects a DS4 twice using USB and BT, we reject the
second device connection after the setup work. We then perform
a cleanup, but during cleanup we are not removing the touchpad
device. This leads to leakage of an input device, which we would
never remove. It can likely result into a kernel oops as well
when the touchpad evdev node is accessed and the underlaying HID
device has been removed from the system.

[jkosina@suse.cz: added stable annotation]
Fixes: ac797b95f532 ("HID: sony: Make the DS4 touchpad a separate device")
Cc: stable@vger.kernel.org
Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit a687c5765b5ae19fe559e14615ddc87ebb46d409)

Bug: 111431828
Signed-off-by: Kim Low <kim-huei.low@sony.com>
Change-Id: Iaf021a5624d96efca291c4890df89cb6ed1550b7
Signed-off-by: Siarhei Vishniakou <svv@google.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>