cpufreq: schedutil: Avoid unnecessary kthread wakeup
We consider the util data for a CPU to be stale if it is older than one
WALT window duration.

However, when the CPUs in a cluster have their util data updated at the end
of a WALT window, some of the active CPUs might get their update a tiny bit
after the end of the WALT window and hence have their util data ignored
when selecting the cluster frequency.

This can momentarily (for a few micro seconds) cause the cluster frequency
to drop to a lower frequency and then come back to the correct frequency.
This momentary drop and the immediate increase in the cluster frequency
causes an unnecessary wake up of the governor's kthread.

This patch avoids this unnecessary wake up by considering the util data for
a CPU to be stale only if it's older than 112% of the WALT window duration.

Change-Id: I931204dad90eb086abb3f684b3a249d60d3c7001
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>