autofs - fix slab out of bounds read in getname_kernel()
Initial patch contributed by Tomas Bortoli.

The autofs subsystem does not check that the "path" parameter is
present for all cases where it is required when it is passed in
via the "param" struct.

In particular it isn't checked for the AUTOFS_DEV_IOCTL_OPENMOUNT_CMD
ioctl command.

To solve it, modify validate_dev_ioctl() function to check that a
path has been provided for ioctl commands that require it.

Bug: 132045014
Change-Id: If204268b0443a50d54735ccc98231424bf1307b8
Signed-off-by: Tomas Bortoli <tomasbortoli@gmail.com>
Signed-off-by: Ian Kent <raven@themaw.net>
Reported-by: syzbot+60c837b428dc84e83a93@syzkaller.appspotmail.com
(cherry picked from commit 8b79d725884ef880173c5030f77b5d6b238dbb02)
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>