UPSTREAM: HID: steam: Fix input device disappearing
The `connected` value for wired devices was not properly initialized,
it must be set to `true` upon creation, because wired devices do not
generate connection events.

When a raw client (the Steam Client) uses the device, the input device
is destroyed. Then, when the raw client finishes, it must be recreated.
But since the `connected` variable was false this never happended.

Signed-off-by: Rodrigo Rivas Costa <rodrigorivascosta@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>

Bug: 147434575
(cherry picked from commit 20eee6e5af35d9586774e80b6e0b1850e7cc9899)
Signed-off-by: Siarhei Vishniakou <svv@google.com>
Change-Id: Ib362e0dec7d53dfc4fcfdca45824cf90358240e2
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>